refactor: extract shared faithfulness scorer with overlap_fraction
